requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
requests
beautifulsoup4
pyjanitor
mysql-connector-python
apache-airflow